+++ packages/windie-sdk-js/cjs/runtime/WindieAgent.js
@@ -1,37 +1,4 @@
 "use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.WindieAgent = void 0;
 const InMemoryConversationStore_js_1 = require("../stores/InMemoryConversationStore.js");
@@ -54,11 +21,7 @@ function logMemoryRetrievalDiagnostic(diagnostic) {
     console.warn(`[Windie SDK] memory retrieval diagnostic: ${details}`);
 }
 class WindieAgent {
-    static async startDesktop(options) {
-        const { WindieDesktopAgent } = await Promise.resolve().then(() => __importStar(require('./WindieDesktopAgent.js')));
-        return WindieDesktopAgent.start(options);
-    }
-    constructor(id, session, agentDefinition, sdkClient, owner, localRuntime, userId = 'local-sdk-user', defaultConversationStore = new InMemoryConversationStore_js_1.InMemoryConversationStore(), memoryEnabled = true) {
+    constructor(id, session, agentDefinition, sdkClient, owner, localRuntime, userId = 'local-sdk-user', defaultConversationStore = new InMemoryConversationStore_js_1.InMemoryConversationStore(), memoryEnabled = true, localToolLifecycle) {
         this.id = id;
         this.session = session;
         this.agentDefinition = agentDefinition;
@@ -68,6 +31,7 @@ class WindieAgent {
         this.userId = userId;
         this.defaultConversationStore = defaultConversationStore;
         this.memoryEnabled = memoryEnabled;
+        this.localToolLifecycle = localToolLifecycle;
     }
     getDefaultConversationStore() {
         return this.defaultConversationStore;
@@ -168,6 +132,9 @@ class WindieAgent {
             sdkClient: this.sdkClient,
             userId: this.userId,
             memoryEnabled: this.memoryEnabled,
+            localToolLifecycle: options.localToolLifecycle === undefined
+                ? this.localToolLifecycle
+                : options.localToolLifecycle,
             enrichQuery: async (input) => {
                 const enriched = await (0, ContextEnrichmentPipeline_js_1.enrichQueryPayload)({
                     text: input.text,